hand-gesture-recognition/requirements.txt
tensorflow>=2.13.0
opencv-python>=4.8.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.1.0
matplotlib>=3.6.0
seaborn>=0.12.0
mlflow>=2.0.0
jupyter>=1.0.0
imageio>=2.31.0
albumentations>=1.3.0  # Replace imgaug with albumentations
Pillow>=9.0.0